move dated_version_panel out of status into its own utility
dated_version_panel is a viewing-context notice ("you are looking at a
historical snapshot; here is the current version"), not a status message.
Moving it out of status.py into lgu2/util/dated_version.py gives it its
own template (new_theme/document/dated_version_panel.html) matching the
extent/in-force/confers-power panel pattern rather than rendering via
status/side_panel.html.

Also moves is_most_recent_version and _shown_milestone_date to the new
module (status.py imports is_most_recent_version back from there), and
consolidates the Link dataclass into util/links.py so both modules share
one definition.

Bonus fix: replaced a multi-line {# #} comment in status/side_panel.html
with {% comment %}...{% endcomment %} — Django's lexer does not strip
multi-line {# #} comments, so the literal text was appearing in rendered
output and causing a test failure.
