+++ lgu2/util/dated_version.py
@@ -1,0 +1,109 @@
+"""Dated-version panel for legislation pages.
+
+The panel shown when a user is viewing a non-current (historical) version.
+This is a viewing-context notice — "you are looking at a snapshot; here is
+how to reach the current version" — not a status message.  It is rendered
+by its own template and is not part of the status bounded context.
+"""
+
+from dataclasses import dataclass
+from datetime import date as _Date
+from typing import List, Optional, Tuple
+
+from django.urls import reverse
+from django.utils.formats import date_format
+from django.utils.translation import gettext as _
+
+from .links import Link, make_document_link, make_fragment_link
+
+
+@dataclass(frozen=True, slots=True)
+class VersionPanel:
+    heading: str
+    paragraphs: Tuple[str, ...] = ()
+    links: Tuple[Link, ...] = ()
+
+
+def is_most_recent_version(meta) -> bool:
+    versions = meta.get('versions') or []
+    return bool(versions) and meta['version'] == versions[-1]
+
+
+def _shown_milestone_date(meta) -> Optional[_Date]:
+    version = meta.get('version', '')
+    try:
+        return _Date.fromisoformat(version)
+    except (TypeError, ValueError):
+        return meta.get('date')
+
+
+def dated_version_panel(meta, lang: Optional[str] = None, section: Optional[str] = None) -> VersionPanel:
+    """Build the panel shown when viewing a non-most-recent version.
+
+    ``lang`` is the route-slug language (``english``/``welsh``), not the
+    API's ``en``/``cy`` code — see ``make_document_link``.
+
+    ``section`` is the fragment href (e.g. ``section/5``) when the panel is
+    being built for a fragment view; the "see most recent version" link
+    then targets the most recent version of that fragment rather than the
+    document as a whole.
+    """
+    title = meta.get('title', '')
+    fragment_info = meta.get('fragmentInfo') or {}
+    fragment_label = fragment_info.get('label') or ''
+    target = (
+        _("{fragment} of {title}").format(fragment=fragment_label, title=title)
+        if fragment_label else title
+    )
+
+    # The "see most recent version" link goes to the document/fragment
+    # route, which accepts regnal years; the "see all changes" link goes
+    # to changes-affected, whose URL pattern is calendar-year-only. So
+    # the two links use different year sources by design.
+    if section:
+        most_recent_href = make_fragment_link(
+            meta['shortType'],
+            meta.get('regnalYear') or meta['year'],
+            meta['number'],
+            section,
+            None,
+            lang,
+        )
+    else:
+        most_recent_href = make_document_link(
+            meta['shortType'],
+            meta.get('regnalYear') or meta['year'],
+            meta['number'],
+            None,
+            lang,
+        )
+
+    links: List[Link] = [Link(
+        text=_("See the most recent version"),
+        href=most_recent_href,
+    )]
+    if meta.get('shortType') and meta.get('year') and meta.get('number'):
+        links.append(Link(
+            text=_("See all changes made to or by {title}").format(title=title),
+            href=reverse(
+                'changes-affected',
+                args=[meta['shortType'], meta['year'], meta['number']],
+            ),
+        ))
+
+    paragraphs: List[str] = [
+        _("This is not the most recent version of {target}.").format(target=target),
+    ]
+    shown_date = _shown_milestone_date(meta)
+    if shown_date:
+        paragraphs.append(
+            _("This legislation has been changed since {date}.").format(
+                date=date_format(shown_date, 'd F Y'),
+            )
+        )
+
+    return VersionPanel(
+        heading=_("Up to date status"),
+        paragraphs=tuple(paragraphs),
+        links=tuple(links),
+    )
